add more tests

+++ part-2/src/Calculator/Constants.cs
@@ -2,9 +2,9 @@
 {
     public static class Constants
     {
-        public static readonly double Pi = 3.14;
-        public static readonly double Root2 = 1.414;
-        public static readonly double Phi = 1.61;
-        public static readonly double e = 2.71;
+        public static readonly decimal Pi = 3.14M;
+        public static readonly decimal Root2 = 1.414M;
+        public static readonly decimal Phi = 1.61M;
+        public static readonly decimal e = 2.71M;
     } 
 }
